improve: add a warning to notify users when all clients are empty
According to Egg.js doc and a known issue
(See:https://github.com/eggjs/egg/pull/3981 ), it seems we don't
allow 'option.client', 'option.clients' and 'option.default' (all are
null). But in order NOT to break down the relased version, a warning is
added there to notify the users.